Temporary patch to help with Chef build
This commit applies some changes that have been found necessary on the
dev branch to ensure that the Chef build proceeds successfully, to the
master branch so that we can then integrate them to prod and staging. In
time, these changes will be overtaken by the changes stacked up on the
dev branch, but we're not ready to merge those yet.