Fix app not running in Rails production mode
FoodStandardsAgency/data-dot-food#43
Fix a problem with missing assets when running the Rails server in
`production` mode. The assets precompile step needs to execute in the
same execution environment as the server execution step, so that the
appropriate `environments/*.rb` config file is applied.

Also optimised the Docker image by omitting some files that are not
needed at run time. Most of these are omitted in the `.dockerfile`,
but the `node_modules` needs to be present for the assets compilation,
so we remove it afterwards.