Fix error when there is no referer
FoodStandardsAgency/data-dot-food#82
If the `Referer` HTTP request field is omitted, the feedback controller
will error and send a Sentry alert. Most likely, these errors are not
user experience issues, because in normal usage the browser will fill-in
the `Referer` field. So the supposition is that these errors are being
caused by crawlers or other bots. Nonetheless, this change will reduce
some Sentry noise.