Fix broken remove-filter link
FoodStandardsAgency/data-dot-food#100
This commit addresses a regression in the "remove filter" action in the
nav bar. Removing a filter was not working, because the `href` being
generated for the action was incorrect. This was traced to the arguments
to `Hash.except()` being passed a list of values, when in fact it uses
varargs.

- add a splat operator to distribute the list of params to `except`
- ensure keys being removed are strings not symbols
- removed some old dead code
- added a test
- ensure that the label on the action is only visible to screen reader
  users