Update Ruby and JS dependencies
A round of general updates for associated Ruby gems and JS packages.
This is partly to address a number of recent CVE warnings.

I did try to update to Webpack 5 as part of these changes, but ran into
problems with webpack itself, wepbacker and PostCSS. So I have left the
webpack dependency at 4.x for the time being.