Fix ISE when trying to render 404
FoodStandardsAgency/data-dot-food#85
If the user enters an incorrect URL for a dataset, they should see a 404
page. A bug in the `render_error` message meant that they would see a
500 Internal Server Error instead.

Having fixed the bug to resolve the ISE, I could see that the error
message exposes the internal API URL for the data catalog API. Since
this is not (yet) published to the public, we don't want to do this. So
I have also added some code to censor the IP address of the API
endpoint. The error message now reads:

```
Not found: Failed to read from
[[SapiNT API]]/catalog/data/dataset/52443157-3f92-409f-8362-4277ee04547c:
404
```

This fix should also serve to reduce the noise we're getting on Sentry
from what should be benign 404 errors.